+++ src/Components/Center/ProductCard.css
@@ -285,7 +285,7 @@ footer .right-icons > * {
   margin: 0px -1px -1px;
 }
 
-.all-products {
+.all-prod_options {
   margin-top: 12px;
 }
